% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_trait_analysis.R
\name{analyze_multiple_traits}
\alias{analyze_multiple_traits}
\title{Multi-trait Analysis Functions}
\usage{
analyze_multiple_traits(data, traits, model_type = "plot_re")
}
\arguments{
\item{data}{Dataset containing trait data}

\item{traits}{Vector of trait names to analyze}

\item{model_type}{Type of model to fit ("plot_re", etc.)}
}
\value{
List of analysis results for each trait
}
\description{
Functions for analyzing multiple traits simultaneously.
Analyze multiple traits
}
\details{
Performs standardized analysis across multiple traits
}
