% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_analysis.R
\name{calculate_single_variogram}
\alias{calculate_single_variogram}
\title{Internal function to calculate variogram for a single phenotype}
\usage{
calculate_single_variogram(data, phenotype_name)
}
\arguments{
\item{data}{Complete-case dataset for the phenotype}

\item{phenotype_name}{Name of the phenotype column}
}
\value{
List containing variogram data and scaling info
}
\description{
Internal function to calculate variogram for a single phenotype
}
\keyword{internal}
