% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variogram_analysis.R
\name{create_directional_variograms}
\alias{create_directional_variograms}
\title{Create directional variograms}
\usage{
create_directional_variograms(data, response, angles = c(0, 45, 90, 135))
}
\arguments{
\item{data}{Spatial data frame}

\item{response}{Response variable name}

\item{angles}{Vector of angles for directional analysis}
}
\value{
Variogram object
}
\description{
Creates directional variograms to assess anisotropy
}
