% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment_effects.R
\name{extract_treatment_effects_emmeans}
\alias{extract_treatment_effects_emmeans}
\title{Extract treatment effects using emmeans}
\usage{
extract_treatment_effects_emmeans(models, best_model_name, phenotype_name)
}
\arguments{
\item{models}{List of fitted models}

\item{best_model_name}{Name of the best model to use}

\item{phenotype_name}{Name of the phenotype}
}
\value{
List containing emmeans results and contrasts
}
\description{
Uses emmeans to extract marginal means and contrasts for treatment effects
}
