% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_models.R
\name{fit_all_models}
\alias{fit_all_models}
\title{Spatial Model Fitting Functions}
\usage{
fit_all_models(data, phenotype_name)
}
\arguments{
\item{data}{Complete-case dataset for the phenotype}

\item{phenotype_name}{Name of the phenotype column}
}
\value{
List of fitted model objects
}
\description{
Functions for fitting hierarchical spatial models and extracting statistics.
Fit all six model structures for a given phenotype
}
\details{
Fits a hierarchy of 6 spatial models from simple fixed effects to 
complex mixed-effects with spatial correlation
}
